OidcClient with login() and getUserManager functions
OidcClient wraps the oidc-client-ts functionalities and adds easier language support.

Also makes it easier to use with React and communicate with middleware.

getUserManager() returns the userManager so it can be called outside.

Currently it is used only in tests.

TestUtil has common functions to use when testing oidcClient directly via React.Context.