Simple signal (event) emitter
OidcClient must be able to send state changes, errors and events to listening functions

This enables devs to show errors when something happens and reactive hooks can be made.

The beacon stores listeners and triggers them.

Syncronous events are sent immediately to all listeners. If a listener triggers another event, it is emitted after all listeners of the first event are triggered.

Asyncronous events are sent one by one. Listener functions are considered asyncscronous.

Next listener is not called until previous promise is fulfilled.

If an async listener triggers async event, that is not sent after first one's listeners are fulfilled.

The first promise is not resolved until all spawned async events are fulfilled.