Hooks for ApiTokensClient
One hook to get the client and one to get the tokens

useApiTokens returns a function to get the tokens and a function the get state of renewal.

Better to return functions than actual values, because values  would not update until re-rendered.

useLoginListener can be used to make a component that re-renders automatically when tokens change.