Added waitForSignals function
The function returns a promise which is resolved, when given signal(s) have been detected. Or it rejects the promise, when a signal in rejection list is given.

The function can be used for racing conditions, for example waiting for a user renewal success or error, which ever comes first.

Or wait for a longer set of emitted signals from user renewal start to api token renewal end.