Added KeyboardTracker
It tracks when given element has focus (via focusin event) or focusout (via focusout). 'focus' and 'blur' are not used, because the tracked element may not be focusable.

When focused the tracker finds element's children to set focus to.  Focusable children as queried with given prop 'childSelector'.

When element has focusin, then keyboard events are tracked and focus is moved accordingly.

Non-visible elements are discarded.

Focus is forced, if element is not focusable by default.