Added onChange and setFocusedElementByIndex functions
Event tracking had some bugs when picking event element. Event.relatedTarget is not the element that is focused.

OnChange communicates the state of the tracker. Useful when items has submenus or focus should be transferred from other elements.

Transferring is revelevant when none of the elements are focusable by default. Tabindex is not set or element is not natively focusable.