Dynamic item updates
Elements may be added or removed at any time. For example when submenus are opened or closed.

Added refresh() function, which forces the tracker to re-build tracking data with a new list of elements.

The tracker it self cannot usually reset focus back where it was, because adding/removing requires user interaction which moves the focus away.

Tracking must be made manually like in the story.