Store focused item only when element is focused
When next/previous is called, the focused item is stored and focus is moved.

If user navigates with mouse or tab or shift+tab, the focus was moved, but not stored properly.

Now next/previous move focus, but new item is stored when focusin-event is triggered.

Now navigation with tab works too without registering the keys to "keys".