MultiLevel tracking and focusing
Keyboard tracker now triggers the onChange callback when its data is updated. This way it is easy to know when re-render is done.

And then focus can be reset easily after new items are added or old ones removed.

This version does not care about the direction of the submenu, so all items are navigated as a one group.